+++ src/UI/Buttons/index.js
@@ -10,7 +10,7 @@ import LogoutButton from './LogoutButton';
 import AddProgressButton from './AddProgressButton';
 import AddUserButton from './AddUserButton';
 import ChangePassButton from './ChangePassButton';
-import HidePassButton from './HidePassButton';
+import PassButton from './PassButton';
 
 export {
   Button,
@@ -25,5 +25,5 @@ export {
   AddProgressButton,
   AddUserButton,
   ChangePassButton,
-  HidePassButton,
+  PassButton,
 };
